feat: Send trip invitations to specified email addresses
This commit adds functionality to send trip invitations to the specified email addresses. The `create` method in the `TripController` now sends invitations to the rest of the travelers using the `sendTripInvitation` method in the `TripService`. The `sendTripInvitation` method emits a `trip.invitation` event with the email, trip details, and current user ID. The `MailService` listens to the `trip.invitation` event and sends an email to the invited user with the trip details.