feat: Add functionality to add and remove participants from a trip
This commit adds the ability to add and remove participants from a trip. Two new PATCH endpoints are introduced in the `TripUserController`: `addParticipant` and `removeParticipant`. These endpoints allow users to add or remove a specific user from a trip by providing the trip ID and user ID as parameters. The corresponding methods `addParticipant` and `removeParticipant` are implemented in the `TripService` to handle the logic of adding and removing participants from the trip. This feature enhances the flexibility and control over trip participants.